commit up ens of module 9

+++ backend/routes/tasks.js
@@ -36,8 +36,10 @@ checkAuth,
   const task = new Task({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
+
   task.save().then(createdTask => {
     res.status(201).json({
       message: 'Task Added Succesfully',
@@ -66,11 +68,15 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: red.userData.userId
   });
-  Task.updateOne({_id: req.params.id }, task ).then(result => {
-    console.log(result);
-    res.status(200).json({message:"update successful"});
+  Task.updateOne({_id: req.params.id, creator:req.userData.userId }, task ).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({message:"Update successful"});
+    } else{
+      res.status(401).json({message:"Not  authorized!"});
+    }
   });
 });
 
@@ -108,11 +114,14 @@ router.get('/:id', (req, res, next) =>{
 });
 
 router.delete('/:id', checkAuth, (req, res, next) =>{
-  Task.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
+  Task.deleteOne({_id: req.params.id, creator:req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      console.log()
+      res.status(200).json({message:"Deletion successful"});
+    } else {
+      res.status(401).json({message:"Not  authorized!"});
+    }
   });
-  res.status(200).json({ message: "Task deleted!"});
-
 });
 
 module.exports = router;
